requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.12
motor==3.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.0
pytest-asyncio==0.21.0
python-dotenv==1.0.0
pymongo==3.12.3
python-multipart
httpx
pytest-mock
